% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_substitutions.R
\name{add_substitutions}
\alias{add_substitutions}
\title{Add dna differences from fasta files to phased tidy vcfR data frame}
\usage{
add_substitutions(HT, Anc, ref, anc)
}
\arguments{
\item{HT}{tidy data of haplotypes, like that coming from vcf_haplos2tidy().}

\item{Anc}{the name of the ancestral sequence in HT}

\item{ref}{path to a fasta (can be gzipped) that contains the "reference" sequence.}

\item{anc}{path to a fasta (can be gzipped) which represents that homologous sequence,
which will often be the "ancstral" sequence which has been aligned to the exact positions
in the reference sequence.}
}
\description{
Here is the situation where you might use this: you have phased sequence
data from chinook salmon, and you also have the ancestral state from coho
for each of those sites.  But now you want to add into that all the sites
where there are differences between coho and Chinook, although Chinook has
no variants at these sites.
}
\details{
Note that a single CHROM is allowed and it is assumed to be the same for HT and ref
and anc.  No checking of this is done at the moment.
}
\examples{
# get the VCF first
V <- system.file(package = "ecaRbioinf", "extdata", "greb1l-imp-phased-with-anc.vcf.gz")
haps <- vcf_haplos2tidy(V, Anc = "Coho")

# whittle it down to the RoSA region
haps$fix <- haps$fix \%>\% dplyr::filter(POS > 12.05e6, POS < 12.4e6)
haps$tidy <- haps$tidy \%>\% dplyr::filter(POS > 12.05e6, POS < 12.4e6)

# get the fasta
ref <- system.file(package = "ecaRbioinf", "extdata", "NC_037124.1_fragment_chinook.fna.gz")
anc <- system.file(package = "ecaRbioinf", "extdata", "NC_037124.1_fragment_coho.fna.gz")

# add the substitutions

combo <- add_substitutions(haps, "Coho", ref, anc)
}
