% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_raster_plot.R
\name{expand_anno_cols}
\alias{expand_anno_cols}
\title{helper function specific to haplo_raster_plot.}
\usage{
expand_anno_cols(A, n)
}
\arguments{
\item{A}{a tibble with hnames, column, value, and width (which is
a proportion of the main raster area that you want the width of the
column to be).}

\item{n}{the width of the main raster area.  Basically the number
of SNPs shown.}
}
\description{
expands the annotation columns into a raster that
can be plotted as another layer
}
