% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_raster_plot.R
\name{expand_anno_rows}
\alias{expand_anno_rows}
\title{helper function specific to haplo_raster_plot.}
\usage{
expand_anno_rows(A, n, start_shift = 0)
}
\arguments{
\item{A}{a tibble with POS, row, value, and height (which is
a proportion of the main raster-area height that you want the height of the
row to be).}

\item{n}{the height of the main raster area.  Basically the number
of hnames shown.}

\item{start_shift}{How many rows down do we want to start these guys?}
}
\description{
expands the annotation rows into a raster that
can be plotted as another layer
}
