% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_sam_flag.R
\name{expand_sam_flag}
\alias{expand_sam_flag}
\title{expand SAM flag into 12 columns}
\usage{
expand_sam_flag(D, column)
}
\arguments{
\item{D}{the tibble of SAM rows}

\item{column}{the name of the column holding the SAM flag}
}
\description{
expand SAM flag into 12 columns
}
