% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_raster_plot.R
\name{haplo_raster_plot}
\alias{haplo_raster_plot}
\title{make a raster plot of allelic types on haplotypes}
\usage{
haplo_raster_plot(D, h_ord = NULL, pos_annot, fcolors,
  highlight_pos = NULL, kinky_highlight_pos = NULL,
  kinky_highlight_color = "red", yaxis_name = "Haplotype",
  xaxis_name = "SNP", pos_bar_height = 0.1, pos_bar_text_size = 1,
  lo_kink_frac = 0.2, hi_kink_frac = 0.2, tick_frac = 0.1,
  expand_top = 0.1, expand_width = 0, kinky_line_size = 0.1,
  het_outline_color = NA, het_outline_size = 1,
  plot_read_depths = FALSE, annotation_columns = NULL,
  annotation_rows = NULL, anno_row_start = 0,
  snp_quant_tibble = NULL, sq_props = list(lo = -9, hi = -1),
  no_legend = FALSE, no_y_labels = FALSE, no_x_labels = TRUE)
}
\arguments{
\item{D}{a tibble with a column \code{POS} that is the positions of the
markers, a column \code{hnames} which are the names of the haplotypes, and
a column \code{atypes} which are the types to be plotted as different colors.
It might also have a column \code{maxGP} which holds the genotype probability
of the called allele.}

\item{h_ord}{if not NULL (the default) this is a vecdtor giving the
order (from top to bottom of
the plot) that you want these to come out.  This will also filter the data set
to include only the hnames in h_ord.  If this is NULL, then the order
of haplotypes in the data set will be used.}

\item{pos_annot}{a tibble with two columns: pname and pos that give the name of the
label you want on the genome position bar, and the position you want it at.}

\item{fcolors}{a named vector to be passed as the values argument to scale_fill_manual. Note that
if you are including more categories in annotation_columns, then you have to include
them, too, for example: fcolors <- c(S = "yellow", F = "blue", spring = "yellow", fall = "blue", Sacto = "red", Klamath = "green")}

\item{highlight_pos}{a tibble with a column named "POS" of SNP positions and column named "name"
which holds the name for SNPs that should be highlighted by surrounding
the columns by a dotted line.}

\item{kinky_highlight_pos}{just like highlight_pos, but operates separately to color
the kinky lines for highlighted positions that overlap with the columns.}

\item{kinky_highlight_color}{By default this is red.}

\item{pos_bar_height}{what fraction of the vertical plot area should we find the position
line above the plot area?}

\item{tick_frac}{the fraction of pos_bar_height that will be half the tick mark height.}

\item{expand_top}{the fraction of pos_bar_height to expand the top of the figure so
that it can capture the text of the positions.}

\item{expand_width}{the fraction of the width of plot to expand by so that it can
capture the text of the positions.}

\item{het_outline_color}{if this is not NA, it is the color which will
be used on the outside of each cell to denote whether it is heterozygous
or not.}

\item{plot_read_depths}{If this is true, this just plots a heat map
identical to the raster haplo that shows read depths.}

\item{annotation_columns}{this is a tibble that has the columns hnames,
column, value, width, and order.  These can be used to show different
characteristics in the negative x range.}

\item{annotation_rows}{this is a tibble that has the columns POS,
row, value, height, and order.  These can be used to show different
characteristics in the negative y range.}

\item{anno_row_start}{which row below the main raster do you want the annotation rows to start?}

\item{kink_frac}{the fraction of the pos_bar_height at which the marker lines should kink.}
}
\description{
This is pretty simple, but it is a standard operation,
and a hassle to hand code over and over again (especially
getting the lines that show the positions of the markers).
}
\examples{
data(RoSA)
data(rosa_9)
D <- RoSA \%>\%
     dplyr::mutate(hnames = haplo_name,
                   atypes = alle2)

# make a tibble to annotate
pos_annot <- tibble::tibble(pos = seq(12.1e6, 12.35e6, by = 0.5e5)) \%>\%
     dplyr::mutate(pname = sprintf("\%.2f", pos/1e06))

# it turns out that these guys are ordered in RoSA how we want them to come out
h_ord <- unique(D$hnames)

# color the "spring" allele yellow and the other blue
fcolors <- c(S = "yellow", F = "blue")

# get positions of SNPs to highlight
h_pos = rosa_9 \%>\% dplyr::select(name, POS) \%>\%
     dplyr::mutate(POS = as.integer(POS))

g <- haplo_raster_plot(D, h_ord, pos_annot, fcolors, h_pos)

# if you want to express genotype posteriors by transparency
# you have to get those data from the unphased but imputed data
U <- system.file(package = "ecaRbioinf", "extdata", "greb1l-imputed-unphased.vcf.gz")
u <- vcfR::read.vcfR(U)
ut <- vcfR::vcfR2tidy(u)$gt \%>\%
     tidyr::separate(gt_GP, into = c("gp0", "gp1", "gp2"), sep = ",", convert = TRUE)
utj <- ut \%>\%
     dplyr::mutate(maxGP = pmax(gp0, gp1, gp2)) \%>\%
     dplyr::select(POS, Indiv, maxGP)

D_GP <- D \%>\%
     dplyr::left_join(utj, by = c("POS", "Indiv"))
}
