% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_variants_into_sequence.R
\name{insert_variants_into_sequences}
\alias{insert_variants_into_sequences}
\title{insert variants into sequences}
\usage{
insert_variants_into_sequences(V, fasta, file = "tmp.txt", lo = NA,
  hi = NA)
}
\arguments{
\item{V}{a tidy data frame like that produced by vcf_haplos2tidy()}

\item{fasta}{path to the fasta file.  Currently it has to have only a
single segment in it, and you should already have extracted the
part of the sequence that you want to propagate the variants into.
This file can be gzipped. The name of the segment has to
have the start position in it (relative to the start of the chromosome
as presumably used in the POS in V)}

\item{lo}{1-based genomic coordinate to start from, if desired.}

\item{hi}{1-based genomic coordinate to end at. Note that lo and hi must be contained
within the fasta.}

\item{variant_extent_overhang}{If this is NULL then the full sequence from
fasta is used.  Otherwise, make it an integer, in which case that is the
number of base pairs from min and max of V$POS that are included in the sequence.}
}
\description{
This assumes a fasta file that is available that includes
all the sequence that you want to propagate the variants
into.
}
\examples{
data(haps)
V <- haps$tidy
fasta <- system.file("extdata", "NC_037124.1_fragment_chinook.fna.gz", package = "ecaRbioinf")
lo <- 12260034
hi <- 12289484
}
