% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_key_val_pairs.R
\name{separate_key_value_pairs}
\alias{separate_key_value_pairs}
\title{separate a column of multiple key=value pairs into separate columns headed by the keys}
\usage{
separate_key_value_pairs(D, column, key_regex = "([a-zA-Z_0-9]+)=",
  val_regex = "=([^;]*)", remove = TRUE)
}
\arguments{
\item{D}{a data frame}

\item{column}{a string-name of the column to separate}

\item{key_regex}{reg-ex to pick out they keys.  Default is for =/;}

\item{val_regex}{reg-ex to pick out the vals. Default is for =/;}

\item{remove}{if TRUE the column is removed}
}
\description{
This is an ugly problem when we have different keys in different rows.  Here I am
going to try an easy-to-code solution that not be super efficient, but interesting
to see, nonetheless.
}
\details{
The slow part is actually just parsing things out with the regexes.  Beyond that
it is all pretty fast.

Note that this does not currently convert columns to different variable types afterward.
That is up to the user.
}
\examples{
data(snp_anno_intersect)
D <- snp_anno_intersect
D2 <- separate_key_value_pairs(D, "biggie")
}
