% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_raster_plot.R
\name{snp_quant_ggplot_calls}
\alias{snp_quant_ggplot_calls}
\title{this is a function to plot quantitative information about each SNP}
\usage{
snp_quant_ggplot_calls(Q, POS_levels, top, bottom,
  background_color = "white", color = "black", rim_color = "black")
}
\arguments{
\item{Q}{the tibble that has POS and value.}

\item{POS_levels}{the levels of the POSes in the current raster plot.  This will
determine the x position when we coerce POS to a factor.}

\item{top}{the row number of the top of the background (corresponds to value = 1)}

\item{bottom}{the row number of the bottom of the background (corresponds to value = 0)}

\item{background_color}{string which is the color desired for the background}

\item{color}{string which is the color desired for the value bar}

\item{rim_color}{color for the line around the box of information}
}
\description{
This bad boy will plot a background tile that covers anything.  It is rimmed with
a line of "rim_color" and it is filled with "background_color" and then it throws down
the SNP-specific information as a height bar of height "value" (from the tibble)
}
\details{
This thing returns a vector of ggplot calls that can be run through eval, and also
the tibble to stuff into it.  (Doing it
this way lets us mix continuous and discrete color scales....)
}
