% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_raster_plot.R
\name{tile_based_column_annotations}
\alias{tile_based_column_annotations}
\title{This function creates layers using geom_tile from *expanded* column_annotations}
\usage{
tile_based_column_annotations(ace, h_ord, fcolors)
}
\arguments{
\item{ace}{the expanded column annotations data frame.  It must have
the columns: hnames, column, value, x.}

\item{h_ord}{how the hnames in D are supposed to be ordered}

\item{fcolors}{a named vector of colors to use for the discrete values}
}
\description{
If fill is mapped to a continuous variable, but you still want your annotations
to have discrete values, this is how you have to do it.
}
\details{
This thing returns a vector of ggplot commands to add the layers
}
