% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_raster_plot.R
\name{tile_based_row_annotations}
\alias{tile_based_row_annotations}
\title{This function creates layers using geom_tile from *expanded* row_annotations}
\usage{
tile_based_row_annotations(are, pos_values, fcolors)
}
\arguments{
\item{are}{the expanded row annotations data frame.  It must have
the columns: POS, row, value, y.}

\item{pos_values}{the levels of the POS values to make them a factor from}

\item{fcolors}{a named vector of colors to use for the discrete values}
}
\description{
If fill is mapped to a continuous variable, but you still want your annotations
to have discrete values, this is how you have to do it.
}
\details{
This thing returns a vector of ggplot commands to add the layers

This assumes that the POS values in are are just the ones that should be there.
}
