% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_haplos2tidy.R
\name{vcf_haplos2tidy}
\alias{vcf_haplos2tidy}
\title{convert phased variants in VCF to tidy data frame of haplotypes}
\usage{
vcf_haplos2tidy(V, Anc = NULL)
}
\arguments{
\item{V}{the path to the VCF.  Note that the allele separator has to be "|" for
everything, as it has to be phased.}

\item{Anc}{if NULL (the default) this is ignored.  If non-null, this is the name
to be given to the ancestral sequence.  If used, this requires that the VCF has
an AA field within the INFO field.}
}
\value{
A list with at least two components: \code{fix}: the fixed infor about the variants
from the VCF, and \code{tidy} a tidy data frame with columns: ChromKey, POS, Indiv,
haplo (a or b), allele.
If Anc is given then that will be a sample with an a-haplo in \code{tidy}.  However
If Anc is given, then the return list will have a third component named
avd, which is just
like tidy, except that it includes an extra column \code{anc_vs_derived} that
has an A if the allele is ancestral (i.e. same as the ancestral sequence), a D
if it is derived (different from the ancestral sequence).
(And also, this data frame does not include the ancestral sequence as an Indiv.)
}
\description{
If the ancestral state is listed this gets put out there too
}
\examples{
V <- system.file(package = "ecaRbioinf", "extdata", "greb1l-imp-phased-with-anc.vcf.gz")
haps <- vcf_haplos2tidy(V, Anc = "Coho")
}
